requirements.txt
appnope
asttokens
attrs
beautifulsoup4
certifi
charset-normalizer
comm
debugpy
decorator
exceptiongroup
executing
h11
idna
importlib_metadata
ipykernel
ipython
jedi
jupyter_client
jupyter_core
matplotlib-inline
nest-asyncio
numpy
outcome
packaging
pandas
parso
pexpect
platformdirs
prompt_toolkit
psutil
psycopg2
ptyprocess
pure_eval
py4j
Pygments
PySocks
pyspark
python-dateutil
pytz
PyYAML
pyzmq
requests
selenium
six
sniffio
sortedcontainers
soupsieve
stack-data
tornado
traitlets
trio
trio-websocket
typing_extensions
tzdata
urllib3
wcwidth
websocket-client
wsproto
zipp